SDL: launch debug window
